LayoutTests/imported/w3c/web-platform-tests/xhr/open-url-encoding-expected.txt


PASS percent encode characters 
FAIL lone surrogate assert_equals: expected "%26%2365533%3B" but got "%26%2355357%3B"
